requirements.txt
# 核心依赖
requests>=2.31.0
pyyaml>=6.0.1
click>=8.1.7
colorama>=0.4.6
psutil>=5.9.6
pandas>=2.1.4
numpy>=1.24.3
matplotlib>=3.7.2
seaborn>=0.12.2

# 数据库相关
sqlalchemy>=2.0.23
sqlite3

# 测试相关
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# 日志和配置
python-dotenv>=1.0.0
structlog>=23.2.0

# 数据处理
openpyxl>=3.1.2
xlrd>=2.0.1
chardet>=5.2.0

# 网络和HTTP
urllib3>=2.1.0
certifi>=2023.11.17

# 日期时间处理
python-dateutil>=2.8.2
chinese-calendar>=1.10.0

# 加密和安全
cryptography>=41.0.7
bcrypt>=4.1.2

# 代码质量
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# 文档生成
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0

# 性能监控
memory-profiler>=0.61.0
line-profiler>=4.1.2

# 异步支持
asyncio
aiohttp>=3.9.1

# 其他工具
tqdm>=4.66.1
tabulate>=0.9.0
rich>=13.7.0
